+++ packages/ui/src/stories/atoms/BackButton.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { userEvent, within } from '@storybook/testing-library'
-import { useTranslation } from 'react-i18next';
+// import { useTranslation } from 'react-i18next';
 // import { useNavigate } from 'react-router-dom';
 import { BackButton } from "../../components/atoms/BackButton";
 // import { Icon } from '../../components/atoms';