+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.3] - 2023-12-25
+
+### Added
+- Migrate BackButton to UI
+- Added BackButton.stories
+- Migrate Icon to UI
+- Added Icon stories
+- Migrate tests to atoms (BackButton and Icon)
+- set i18n within StoryBook
+- Migrate languages locales to UI
+- Added 'storybook-react-i18next' to addons in main.ts,
+- Updated preview with globals languages
+
 ## [0.2.1] - 2023-12-19
 
 ### Fix